+++ src/lib.rs
@@ -1,5 +1,4 @@
 pub mod command;
 pub mod error;
-mod utils;
-pub use utils::{load_data_store, parse_request};
 pub mod execution_result;
+pub mod utils;
